+++ app/models/export/activity_forecast_columns.rb
@@ -25,6 +25,10 @@ class Export::ActivityForecastColumns
     }.to_h
   end
 
+  def rows_for_first_financial_quarter
+    rows.each_with_object({}) { |(key, values), rows| rows[key] = values.first }
+  end
+
   private
 
   def forecast_data(activity)
@@ -48,9 +52,16 @@ class Export::ActivityForecastColumns
   end
 
   def forecasts_to_hash
-    @_forecasts_to_hash ||= forecasts.each_with_object({}) { |forecast, hash|
-      hash[[forecast.parent_activity_id, forecast.financial_quarter, forecast.financial_year]] = forecast.value
-    }
+    @_forecasts_to_hash ||=
+      forecasts.each_with_object({}) { |forecast, hash|
+        hash[
+          [
+            forecast.parent_activity_id,
+            forecast.financial_quarter,
+            forecast.financial_year,
+          ]
+        ] = forecast.value
+      }
   end
 
   def all_financial_quarters_with_forecasts