+++ app/services/export/spending_breakdown.rb
@@ -18,24 +18,30 @@ class Export::SpendingBreakdown
       Export::ActivityDeliveryPartnerOrganisationColumn.new(activities_relation: @activities)
     @actual_columns =
       Export::ActivityActualsColumns.new(activities: @activities, include_breakdown: true)
+    @forecast_columns =
+      Export::ActivityForecastColumns.new(activities: @activities, starting_financial_quarter: first_forecast_financial_quarter)
   end
 
   def headers
-    return @activity_attributes.headers if @actual_columns.rows.empty? && forecasts.empty?
+    return @activity_attributes.headers if actuals_rows.empty? && forecast_rows.empty?
 
     @activity_attributes.headers +
       @delivery_partner_organisations.headers +
       @actual_columns.headers +
-      forecasts_headers
+      @forecast_columns.headers
   end
 
   def rows
-    return [] if @actual_columns.rows.empty? && forecasts.empty?
+    return [] if actuals_rows.empty? && forecast_rows.empty?
+
+    attribute_row_data = @activity_attributes.rows
+    delivery_partner_organisations_row_data = @delivery_partner_organisations.rows
+
     activities.map do |activity|
-      @activity_attributes.rows.fetch(activity.id, nil) +
-        @delivery_partner_organisations.rows.fetch(activity.id, nil) +
-        @actual_columns.rows.fetch(activity.id, nil) +
-        forecast_data(activity)
+      attribute_row_data.fetch(activity.id, nil) +
+        delivery_partner_organisations_row_data.fetch(activity.id, nil) +
+        actuals_rows.fetch(activity.id, nil) +
+        forecast_rows.fetch(activity.id, nil)
     end
   end
 
@@ -49,14 +55,17 @@ class Export::SpendingBreakdown
 
   private
 
-  def forecast_data(activity)
-    all_forecast_financial_quarter_range.map { |fq| forecasts_to_hash.fetch([activity.id, fq.quarter, fq.financial_year.start_year], 0) }
+  def actuals_rows
+    @_actuals_rows ||= @actual_columns.rows
   end
 
-  def forecasts_to_hash
-    @_forecasts_to_hash ||= forecasts.each_with_object({}) { |forecast, hash|
-      hash[[forecast.parent_activity_id, forecast.financial_quarter, forecast.financial_year]] = forecast.value
-    }
+  def forecast_rows
+    @_forecast_rows ||= @forecast_columns.rows
+  end
+
+  def first_forecast_financial_quarter
+    return nil if actuals_rows.empty?
+    @actual_columns.last_financial_quarter.succ
   end
 
   def activities
@@ -65,34 +74,6 @@ class Export::SpendingBreakdown
     else
       Activity.includes(:organisation).where(organisation_id: @organisation.id, source_fund_code: @source_fund.id)
         .or(Activity.includes(:organisation).where(extending_organisation_id: @organisation.id, source_fund_code: @source_fund.id))
-    end
-  end
-
-  def forecasts
-    overview = ForecastOverview.new(activity_ids)
-    @_forecasts ||= overview.latest_values
-  end
-
-  def activity_ids
-    activities.pluck(:id)
-  end
-
-  def all_financial_quarters_with_forecasts
-    return [] unless forecasts.present?
-    forecasts.map(&:own_financial_quarter).uniq
-  end
-
-  def forecasts_headers
-    all_forecast_financial_quarter_range.map do |financial_quarter|
-      "Forecast #{financial_quarter}"
-    end
-  end
-
-  def all_forecast_financial_quarter_range
-    @_forecast_quarter_range ||= begin
-      return [] if all_financial_quarters_with_forecasts.blank?
-
-      Range.new(@actual_columns.last_financial_quarter.succ, all_financial_quarters_with_forecasts.max)
     end
   end
 end