+++ spec/services/export/spending_breakdown_spec.rb
@@ -10,51 +10,43 @@ RSpec.describe Export::SpendingBreakdown do
     q1_2019_report = create(
       :report,
       :approved,
-      organisation: @organisation,
+      organisation: @activity.organisation,
       fund: @activity.associated_fund,
       financial_quarter: 1,
       financial_year: 2019
     )
-
-    ForecastHistory
-      .new(@activity, report: q1_2019_report, financial_quarter: 1, financial_year: 2020)
-      .set_value(10_000)
-    ForecastHistory
-      .new(@activity, report: q1_2019_report, financial_quarter: 4, financial_year: 2020)
-      .set_value(5_00)
-    ForecastHistory
-      .new(@activity, report: q1_2019_report, financial_quarter: 1, financial_year: 2021)
-      .set_value(10_000)
-    ForecastHistory
-      .new(@activity, report: q1_2019_report, financial_quarter: 4, financial_year: 2021)
-      .set_value(20_000)
+    forecasts_for_report_from_table(q1_2019_report,
+      <<~TABLE
+        |financial_quarter|financial_year|value|
+        |1                |2020          |10000|
+        |4                |2020          |  500|
+        |1                |2021          |10000|
+        |4                |2021          |20000|
+      TABLE
+    )
 
     q4_2019_report = create(
       :report,
       :approved,
-      organisation: @organisation,
+      organisation: @activity.organisation,
       fund: @activity.associated_fund,
       financial_quarter: 4,
       financial_year: 2019
     )
-
-    ForecastHistory
-      .new(@activity, report: q4_2019_report, financial_quarter: 1, financial_year: 2020)
-      .set_value(5_000)
-    ForecastHistory
-      .new(@activity, report: q4_2019_report, financial_quarter: 4, financial_year: 2020)
-      .set_value(2_500)
-    ForecastHistory
-      .new(@activity, report: q4_2019_report, financial_quarter: 1, financial_year: 2021)
-      .set_value(20_000)
-    ForecastHistory
-      .new(@activity, report: q4_2019_report, financial_quarter: 4, financial_year: 2021)
-      .set_value(10_000)
+    forecasts_for_report_from_table(q4_2019_report,
+      <<~TABLE
+        |financial_quarter|financial_year|value|
+        |1                |2020          | 5000|
+        |4                |2020          | 2500|
+        |1                |2021          |20000|
+        |4                |2021          |10000|
+      TABLE
+    )
 
     @q1_report = create(:report, financial_quarter: 1, financial_year: 2020)
     @q2_report = create(:report, financial_quarter: 2, financial_year: 2020)
 
-    create_fixtures(
+    actuals_from_table(
       <<~TABLE
         |transaction|report|financial_period|value|
         | Actual    |q1    | q1             |  100|
@@ -185,6 +177,40 @@ RSpec.describe Export::SpendingBreakdown do
       expect(value_for_header("Forecast FQ3 2021-2022")).to eq 0
     end
 
+    it "attibute rows are only create once" do
+      rows_data_double = double(Hash, fetch: [], empty?: false)
+
+      attribute_double = double(rows: rows_data_double)
+      allow(Export::ActivityAttributesColumns).to receive(:new).and_return(attribute_double)
+
+      delivery_partner_organisation_double = double(rows: rows_data_double)
+      allow(Export::ActivityDeliveryPartnerOrganisationColumn).to receive(:new).and_return(delivery_partner_organisation_double)
+
+      actuals_double = double(rows: rows_data_double, last_financial_quarter: FinancialQuarter.new(2, 2021))
+      allow(Export::ActivityActualsColumns).to receive(:new).and_return(actuals_double)
+
+      forecasts_double = double(rows: rows_data_double)
+      allow(Export::ActivityForecastColumns).to receive(:new).and_return(forecasts_double)
+
+      subject.rows
+
+      expect(attribute_double)
+        .to have_received(:rows)
+        .once
+
+      expect(delivery_partner_organisation_double)
+        .to have_received(:rows)
+        .once
+
+      expect(actuals_double)
+        .to have_received(:rows)
+        .once
+
+      expect(forecasts_double)
+        .to have_received(:rows)
+        .once
+    end
+
     context "where there are additional activities" do
       before do
         create_list(:project_activity, 4, organisation: @organisation)
@@ -213,7 +239,7 @@ RSpec.describe Export::SpendingBreakdown do
     end
   end
 
-  def create_fixtures(table)
+  def actuals_from_table(table)
     CSV.parse(table, col_sep: "|", headers: true).each do |row|
       case row["transaction"].strip
       when "Actual"
@@ -227,6 +253,17 @@ RSpec.describe Export::SpendingBreakdown do
       else
         raise "don't know what to do"
       end
+    end
+  end
+
+  def forecasts_for_report_from_table(report, table)
+    CSV.parse(table, col_sep: "|", headers: true).each do |row|
+      ForecastHistory.new(
+        @activity,
+        report: report,
+        financial_quarter: row["financial_quarter"].to_i,
+        financial_year: row["financial_year"].to_i,
+      ).set_value(row["value"].to_i)
     end
   end